pyproject.toml
[project]
name = "toolflood"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "datasets==4.4.2",
    "faiss-cpu==1.13.2",
    "numpy==2.4.0",
    "pyarrow==22.0.0",
]

[dependency-groups]
dev = [
    "datasets==4.4.2",
    "fsspec==2025.10.0",
    "huggingface-hub==0.36.0",
    "ipykernel==7.1.0",
    "matplotlib==3.10.8",
    "seaborn==0.13.2",
    "langchain==1.2.0",
    "langchain-azure-ai==1.0.0",
    "langchain-community==0.4.1",
    "langchain-huggingface==1.2.0",
    "langchain-openai==1.1.6",
    "langchain-xai==1.2.0",
    "loguru==0.7.3",
    "nltk==3.9.2",
    "openpyxl==3.1.5",
    "pandas==2.3.3",
    "pydantic==2.12.5",
    "pyyaml==6.0.3",
    "rouge-score==0.1.2",
    "sentence-transformers==5.2.0",
    "torch==2.9.1",
    "tqdm==4.67.1",
]
